requirements/base.txt
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
mysqlclient==2.1.0
celery==5.2.1
redis==4.0.2
flower==1.0.0
django-cors-headers==3.10.1
djangorestframework-simplejwt==5.0.0
